AndCountReduce kernel, reduce malloc, no sampling
Consolidated kernels into AndCountReduce to reduce kernel launches
Moved malloc/free out from build_tree to build_forest
Removed the scrambling and stratified sampling steps from cubrif.py

+++ brif.h
@@ -45,7 +45,7 @@ typedef int INT;
 // Declarations of this file
 
 typedef double numeric_t;
-typedef int64_t integer_t;
+typedef int integer_t;
 typedef unsigned int bitblock_t;
 
 typedef struct ycode {
